chore: config chain in separate file
