+++ src/iam/models.ts
@@ -1,0 +1,7 @@
+export interface ChainConfig {
+  rpcUrl: string,
+  ensRegistryAddress: string,
+  ensResolverAddress: string,
+  didContractAddress: string,
+  cacheServerUrl: string;
+}
